FoX IO errors initialized at startup
FoX libraries initialized the IO error code everytime it opened a file
for reading. For doing this it needed to open and delete a temporary scratch file.
This operation becomes quite annoying if done by  many thousand of MPI
at the same time. This commit introduces a slightly modified version of
the FoX libs. Error codes are initialized only once and only on  image ionodes.
Codes are then broadcasted to all nodes and stored in memory.