Added IOSTAT check when opening temp UPF files
Added a check on IOSTAT when opening UPF files if this not possible
to open the temp UPF file for some node the program stops after
printing an hopefully explicative error message.

Added also a small md file with the instructions needed to fix UPF
files in order to make them XML syntac compliant.